Индивидуальное 3

+++ LibraryCPP/vector.h
@@ -1,9 +1,11 @@
+#include <stdlib.h>
+
 #ifndef VECTOR_H
 #define VECTOR_H
 
 // Vector (dynamic array)
 // Stores integer values inside
-typedef int Data;
+typedef size_t Data;
 
 struct Vector;
